requirement.txt
alembic==1.3.1
aniso8601==8.0.0
blinker==1.4
Click==7.0
Flask==1.1.1
Flask-JWT-Extended==3.24.1
Flask-Mail==0.9.1
flask-marshmallow==0.8.0
Flask-Migrate==2.1.1
Flask-RESTful==0.3.6
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.10.3
Mako==1.1.0
MarkupSafe==1.1.1
marshmallow==2.14.0
marshmallow-sqlalchemy==0.13.2
pycrypto==2.6.1
PyJWT==1.7.1
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2019.3
six==1.13.0
SQLAlchemy==1.3.11
Werkzeug==0.16.0
flask-cors==3.0.8
